+++ package.json
@@ -73,6 +73,7 @@
     "react-router-dom": "^5.2.0",
     "react-slick": "^0.27.4",
     "redux": "^4.0.5",
-    "slick-carousel": "^1.8.1"
+    "slick-carousel": "^1.8.1",
+    "uuid": "^8.3.1"
   }
 }
