+++ src/components/container/WriteFormContainer.jsx
@@ -1,7 +1,10 @@
-import React from 'react';
+import React, { useState, useEffect } from 'react';
 
 import { useDispatch, useSelector } from 'react-redux';
+
 import WriteForm from '../presentational/WriteForm';
+import ImagesDropzone from '../presentational/ImagesDropzone';
+import ImagePreview from '../presentational/ImagePreview';
 
 import {
   writeNewProduct,
@@ -9,6 +12,8 @@ import {
 } from '../../productSlice';
 
 export default function WriteFormContainer() {
+  const [files, setFiles] = useState([]);
+
   const dispatch = useDispatch();
 
   const newProduct = useSelector((state) => state.productReducer.newProduct);
@@ -18,14 +23,33 @@ export default function WriteFormContainer() {
   }
 
   function handleSubmit() {
-    dispatch(postProduct());
+    dispatch(postProduct({ files }));
+    setFiles([]);
   }
 
+  function handleOnDrop(acceptedFiles) {
+    setFiles([
+      ...files,
+      ...acceptedFiles.map((file) => Object.assign(file, {
+        preview: URL.createObjectURL(file),
+      })),
+    ]);
+  }
+
+  useEffect(() => () => {
+    // Make sure to revoke the data uris to avoid memory leaks
+    files.forEach((file) => URL.revokeObjectURL(file.preview));
+  }, [files]);
+
   return (
-    <WriteForm
-      newProduct={newProduct}
-      onChange={handleChange}
-      onSubmit={handleSubmit}
-    />
+    <div>
+      <ImagesDropzone onDrop={handleOnDrop} />
+      <ImagePreview files={files} />
+      <WriteForm
+        newProduct={newProduct}
+        onChange={handleChange}
+        onSubmit={handleSubmit}
+      />
+    </div>
   );
 }
